Placed all breaking into current
